Inicio da implementação das classes para trabalho

+++ Trabalho/aula_abstract/src/Biologo.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class Biologo extends Funcionario{
 	//Construtor
 	public Biologo(String nome,String sobrenome, int cpf, String email, Date nascimento,
-			int salario, int carga_horaria, long telefone, Time horaDeChegar,
+			double salario, int carga_horaria, long telefone, Time horaDeChegar,
 			Time horaDeSair) {
 		super(nome,sobrenome, cpf, email, nascimento, salario, carga_horaria, telefone,
 				horaDeChegar, horaDeSair);
