+++ Trabalho/aula_abstract/src/Funcionario.java
@@ -3,7 +3,7 @@ import java.util.Date;
 
 public abstract class Funcionario extends Pessoa{
 	//Atributos
-	int salario;
+	double salario;
 	int carga_horaria;
 	long telefone;
 	Time horaDeChegar;
@@ -11,7 +11,7 @@ public abstract class Funcionario extends Pessoa{
 
 	//Construtor
 	public Funcionario(String nome,String sobrenome, int cpf, String email, Date nascimento,
-			int salario, int carga_horaria, long telefone, Time horaDeChegar,
+			double salario, int carga_horaria, long telefone, Time horaDeChegar,
 			Time horaDeSair) {
 		super(nome,sobrenome, cpf, email, nascimento);
 		this.salario = salario;
