+++ Trabalho/aula_abstract/src/Tratador.java
@@ -4,7 +4,7 @@ import java.util.Date;
 public class Tratador extends Funcionario {
 
 	public Tratador(String nome, String sobrenome,int cpf, String email, Date nascimento,
-			int salario, int carga_horaria, long telefone, Time horaDeChegar,
+			double salario, int carga_horaria, long telefone, Time horaDeChegar,
 			Time horaDeSair) {
 		super(nome, sobrenome,cpf, email, nascimento, salario, carga_horaria, telefone,
 				horaDeChegar, horaDeSair);
